+++ src/main/java/com/alper/worldcup/service/LeaderboardTickerEntry.java
@@ -1,4 +1,4 @@
 package com.alper.worldcup.service;
 
-public record LeaderboardTickerEntry(int rank, String username, String displayName, long points) {
+public record LeaderboardTickerEntry(int rank, String username, String displayName, double points) {
 }
