+++ src/main/java/com/alper/worldcup/service/PointsService.java
@@ -5,15 +5,15 @@ import com.alper.worldcup.entity.Team;
 
 public interface PointsService {
 
-    int calculatePoints(int guessHome, int guessAway, int actualHome, int actualAway, MatchStage stage);
+    double calculatePoints(int guessHome, int guessAway, int actualHome, int actualAway, MatchStage stage);
 
-    int calculateKnockoutPoints(int guessHome,
-                                int guessAway,
-                                int actualHome,
-                                int actualAway,
-                                MatchStage stage,
-                                Boolean penaltyGuess,
-                                Boolean penaltyActual,
-                                Team advancingGuess,
-                                Team advancingActual);
+    double calculateKnockoutPoints(int guessHome,
+                                   int guessAway,
+                                   int actualHome,
+                                   int actualAway,
+                                   MatchStage stage,
+                                   Boolean penaltyGuess,
+                                   Boolean penaltyActual,
+                                   Team advancingGuess,
+                                   Team advancingActual);
 }
